requirements.txt
pyyaml
prompt_toolkit